HRWSI_System/requirements.txt
numpy==1.26.2
matplotlib==3.8.1
seaborn==0.13.0
pyyaml==6.0.1
psycopg2-binary==2.9.9
requests==2.28.1
urllib3==1.26.18
pytest-mock==3.12.0
nest_asyncio==1.5.8
kombu==5.3.5